# generated/order_service_contracts/http/openapi.gen.go
// Package orderhttp provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.6.0 DO NOT EDIT.
package orderhttp

import (
	"fmt"
	"net/http"

	"github.com/go-chi/chi/v5"
	"github.com/oapi-codegen/runtime"
	openapi_types "github.com/oapi-codegen/runtime/types"
)

// Defines values for OrderStatus.
const (
	CANCELLED      OrderStatus = "CANCELLED"
	PAID           OrderStatus = "PAID"
	PENDINGPAYMENT OrderStatus = "PENDING_PAYMENT"
)

// Valid indicates whether the value is a known member of the OrderStatus enum.
func (e OrderStatus) Valid() bool {
	switch e {
	case CANCELLED:
		return true
	case PAID:
		return true
	case PENDINGPAYMENT:
		return true
	default:
		return false
	}
}

// Defines values for PaymentMethod.
const (
	CARD          PaymentMethod = "CARD"
	CREDITCARD    PaymentMethod = "CREDIT_CARD"
	INVESTORMONEY PaymentMethod = "INVESTOR_MONEY"
	SBP           PaymentMethod = "SBP"
	UNKNOWN       PaymentMethod = "UNKNOWN"
)

// Valid indicates whether the value is a known member of the PaymentMethod enum.
func (e PaymentMethod) Valid() bool {
	switch e {
	case CARD:
		return true
	case CREDITCARD:
		return true
	case INVESTORMONEY:
		return true
	case SBP:
		return true
	case UNKNOWN:
		return true
	default:
		return false
	}
}

// CreateOrderRequest defines model for CreateOrderRequest.
type CreateOrderRequest struct {
	// PartUuids Список UUID деталей
	PartUuids []openapi_types.UUID `json:"part_uuids"`

	// UserUuid UUID пользователя
	UserUuid openapi_types.UUID `json:"user_uuid"`
}

// CreateOrderResponse defines model for CreateOrderResponse.
type CreateOrderResponse struct {
	// OrderUuid UUID созданного заказа
	OrderUuid openapi_types.UUID `json:"order_uuid"`

	// TotalPrice Итоговая стоимость заказа
	TotalPrice float64 `json:"total_price"`
}

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	// Message Описание ошибки
	Message string `json:"message"`
}

// Order defines model for Order.
type Order struct {
	// OrderUuid Уникальный идентификатор заказа
	OrderUuid openapi_types.UUID `json:"order_uuid"`

	// PartUuids Список UUID деталей
	PartUuids []openapi_types.UUID `json:"part_uuids"`

	// PaymentMethod Способ оплаты (заполняется после оплаты)
	PaymentMethod *PaymentMethod `json:"payment_method,omitempty"`

	// Status Статус заказа:
	// - `PENDING_PAYMENT` — ожидает оплаты
	// - `PAID` — оплачен
	// - `CANCELLED` — отменён
	Status OrderStatus `json:"status"`

	// TotalPrice Итоговая стоимость заказа
	TotalPrice float64 `json:"total_price"`

	// TransactionUuid UUID транзакции (заполняется после оплаты)
	TransactionUuid *openapi_types.UUID `json:"transaction_uuid,omitempty"`

	// UserUuid UUID пользователя
	UserUuid openapi_types.UUID `json:"user_uuid"`
}

// OrderStatus Статус заказа:
// - `PENDING_PAYMENT` — ожидает оплаты
// - `PAID` — оплачен
// - `CANCELLED` — отменён
type OrderStatus string

// PayOrderRequest defines model for PayOrderRequest.
type PayOrderRequest struct {
	// PaymentMethod Способ оплаты:
	// - `UNKNOWN` — неизвестный
	// - `CARD` — банковская карта
	// - `SBP` — система быстрых платежей
	// - `CREDIT_CARD` — кредитная карта
	// - `INVESTOR_MONEY` — деньги инвестора
	PaymentMethod PaymentMethod `json:"payment_method"`
}

// PayOrderResponse defines model for PayOrderResponse.
type PayOrderResponse struct {
	// TransactionUuid UUID транзакции оплаты
	TransactionUuid openapi_types.UUID `json:"transaction_uuid"`
}

// PaymentMethod Способ оплаты:
// - `UNKNOWN` — неизвестный
// - `CARD` — банковская карта
// - `SBP` — система быстрых платежей
// - `CREDIT_CARD` — кредитная карта
// - `INVESTOR_MONEY` — деньги инвестора
type PaymentMethod string

// CreateOrderJSONRequestBody defines body for CreateOrder for application/json ContentType.
type CreateOrderJSONRequestBody = CreateOrderRequest

// PayOrderJSONRequestBody defines body for PayOrder for application/json ContentType.
type PayOrderJSONRequestBody = PayOrderRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Создать заказ
	// (POST /orders)
	CreateOrder(w http.ResponseWriter, r *http.Request)
	// Получить заказ по UUID
	// (GET /orders/{order_uuid})
	GetOrder(w http.ResponseWriter, r *http.Request, orderUuid openapi_types.UUID)
	// Отменить заказ
	// (POST /orders/{order_uuid}/cancel)
	CancelOrder(w http.ResponseWriter, r *http.Request, orderUuid openapi_types.UUID)
	// Оплатить заказ
	// (POST /orders/{order_uuid}/pay)
	PayOrder(w http.ResponseWriter, r *http.Request, orderUuid openapi_types.UUID)
}

// Unimplemented server implementation that returns http.StatusNotImplemented for each endpoint.

type Unimplemented struct{}

// Создать заказ
// (POST /orders)
func (_ Unimplemented) CreateOrder(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Получить заказ по UUID
// (GET /orders/{order_uuid})
func (_ Unimplemented) GetOrder(w http.ResponseWriter, r *http.Request, orderUuid openapi_types.UUID) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Отменить заказ
// (POST /orders/{order_uuid}/cancel)
func (_ Unimplemented) CancelOrder(w http.ResponseWriter, r *http.Request, orderUuid openapi_types.UUID) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Оплатить заказ
// (POST /orders/{order_uuid}/pay)
func (_ Unimplemented) PayOrder(w http.ResponseWriter, r *http.Request, orderUuid openapi_types.UUID) {
	w.WriteHeader(http.StatusNotImplemented)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// CreateOrder operation middleware
func (siw *ServerInterfaceWrapper) CreateOrder(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.CreateOrder(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetOrder operation middleware
func (siw *ServerInterfaceWrapper) GetOrder(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "order_uuid" -------------
	var orderUuid openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "order_uuid", chi.URLParam(r, "order_uuid"), &orderUuid, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "uuid"})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "order_uuid", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetOrder(w, r, orderUuid)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// CancelOrder operation middleware
func (siw *ServerInterfaceWrapper) CancelOrder(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "order_uuid" -------------
	var orderUuid openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "order_uuid", chi.URLParam(r, "order_uuid"), &orderUuid, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "uuid"})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "order_uuid", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.CancelOrder(w, r, orderUuid)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// PayOrder operation middleware
func (siw *ServerInterfaceWrapper) PayOrder(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "order_uuid" -------------
	var orderUuid openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "order_uuid", chi.URLParam(r, "order_uuid"), &orderUuid, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "uuid"})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "order_uuid", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PayOrder(w, r, orderUuid)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{})
}

type ChiServerOptions struct {
	BaseURL          string
	BaseRouter       chi.Router
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, r chi.Router) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseRouter: r,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, r chi.Router, baseURL string) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseURL:    baseURL,
		BaseRouter: r,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options ChiServerOptions) http.Handler {
	r := options.BaseRouter

	if r == nil {
		r = chi.NewRouter()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/orders", wrapper.CreateOrder)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/orders/{order_uuid}", wrapper.GetOrder)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/orders/{order_uuid}/cancel", wrapper.CancelOrder)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/orders/{order_uuid}/pay", wrapper.PayOrder)
	})

	return r
}